requirements.txt
beautifulsoup4==4.12.3
flask==3.0.2
networkx==3.3
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
psycopg2-binary==2.9.10
requests==2.32.3
scikit-learn==1.5.1
tqdm==4.66.5
warcio==1.7.5 